+++ app/javascript/Components/__tests__/submission_table.test.jsx
@@ -95,6 +95,89 @@ describe("For the SubmissionTable's display of inactive groups", () => {
   });
 });
 
+describe("For the SubmissionTable's display of assigned submissions", () => {
+  beforeEach(async () => {
+    const groups_sample = [
+      {
+        _id: 1,
+        max_mark: 21.0,
+        group_name: "assigned_group",
+        tags: [],
+        marking_state: "released",
+        submission_time: "Monday, March 25, 2024, 01:49:14 PM EDT",
+        result_id: 1,
+        final_grade: 12.0,
+        members: [["c6scriab", false]],
+        grace_credits_used: 1,
+        assigned: true,
+      },
+      {
+        _id: 2,
+        max_mark: 21.0,
+        group_name: "unassigned_group",
+        tags: [],
+        marking_state: "released",
+        submission_time: "Monday, March 25, 2024, 01:49:14 PM EDT",
+        result_id: 2,
+        final_grade: 7.0,
+        members: [["g8butter", false]],
+        grace_credits_used: 1,
+        assigned: false,
+      },
+    ];
+    fetch.mockReset();
+    fetch.mockResolvedValueOnce({
+      ok: true,
+      json: jest.fn().mockResolvedValueOnce({
+        groupings: groups_sample,
+        sections: {},
+      }),
+    });
+
+    render(
+      <SubmissionTable
+        assignment_id={1}
+        course_id={1}
+        show_grace_tokens={true}
+        show_sections={true}
+        is_timed={false}
+        is_scanned_exam={false}
+        release_with_urls={false}
+        can_collect={true}
+        can_run_tests={false}
+        can_view_assigned_submissions_only={true}
+        initial_show_assigned_submissions_only={true}
+        defaultFiltered={[{id: "", value: ""}]}
+      />
+    );
+    await screen.findByText("assigned_group");
+  });
+
+  it("contains the correct amount of assigned submissions in the hidden tooltip", () => {
+    expect(
+      screen.getByTestId("show_assigned_submissions_only_tooltip").getAttribute("title")
+    ).toEqual("1 assigned submission");
+  });
+
+  it("initially displays only assigned submissions", () => {
+    expect(screen.getByText("assigned_group")).toBeInTheDocument();
+    expect(screen.queryByText("unassigned_group")).not.toBeInTheDocument();
+  });
+
+  it("displays all submissions after a single toggle", () => {
+    fireEvent.click(screen.getByTestId("show_assigned_submissions_only"));
+    expect(screen.getByText("assigned_group")).toBeInTheDocument();
+    expect(screen.getByText("unassigned_group")).toBeInTheDocument();
+  });
+
+  it("only displays assigned submissions after two toggles", () => {
+    fireEvent.click(screen.getByTestId("show_assigned_submissions_only"));
+    fireEvent.click(screen.getByTestId("show_assigned_submissions_only"));
+    expect(screen.getByText("assigned_group")).toBeInTheDocument();
+    expect(screen.queryByText("unassigned_group")).not.toBeInTheDocument();
+  });
+});
+
 describe("For the SubmissionTable's group name search", () => {
   beforeEach(async () => {
     const groups_sample = [